adding flag to turn off hit recalibration (MuonTrackSteeringTools-01-01-19)
	* flag to turn off hit recalibration for cosmics (ATLASRECTS-1871)
	* fixing compiler warning
	* tagging MuonTrackSteeringTools-01-01-19

2015-03-03 Edward.Moyse@cern.ch
	* Fix for bug ATLASRECTS-1767
	* MuonTrackSteeringTools-01-01-18

+++ MuonSpectrometer/MuonReconstruction/MuonTrackMakers/MuonTrackMakerTools/MuonTrackSteeringTools/src/MooTrackFitter.cxx
@@ -2406,7 +2406,7 @@ namespace Muon {
       for( TrkDriftCircleMath::DCOnTrackCit dcit = matchedDCs.begin();dcit!=matchedDCs.end();++dcit ){
         
         if( dcit->state() == TrkDriftCircleMath::DCOnTrack::OnTrack ) continue;
-	if(dcit->index()>=indexIdMap.size()) continue;
+	if( (unsigned int) dcit->index() >= indexIdMap.size() ) continue;
         indexIdMap[dcit->index()].second = true;
       }
